Move DataTables unit tests to their own test project
