+++ Testing/dotNetRdf.Data.DataTables.Tests/dotNetRdf.Data.DataTables.Tests.csproj
@@ -2,7 +2,7 @@
 
   <PropertyGroup>
     <TargetFrameworks>net472;netcoreapp3.1</TargetFrameworks>
-    <RootNamespace>VDS.RDF</RootNamespace>
+
     <IsPackable>false</IsPackable>
   </PropertyGroup>
 
@@ -20,12 +20,23 @@
   </ItemGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\..\Libraries\dotNetRdf.Inferencing\dotNetRdf.Inferencing.csproj" />
+    <ProjectReference Include="..\..\Libraries\dotNetRdf.Data.DataTables\dotNetRdf.Data.DataTables.csproj" />
+    <ProjectReference Include="..\..\Libraries\dotNetRdf.Sparql\dotNetRdf.Sparql.csproj" />
     <ProjectReference Include="..\dotNetRdf.TestSupport\dotNetRdf.TestSupport.csproj" />
   </ItemGroup>
 
+  <ItemGroup Condition=" '$(TargetFramework)' == 'net472' ">
+    <Reference Include="System.Runtime.Serialization" />
+    <Reference Include="System.Data" />
+    <Reference Include="System" />
+    <Reference Include="Microsoft.CSharp" />
+    <Reference Include="System.Configuration" />
+    <Reference Include="System.Web" />
+    <Reference Include="System.Net.Http" />
+  </ItemGroup>
+
   <ItemGroup>
-    <None Update="resources\InferenceTest.ttl">
+    <None Update="resources\**\*.*;resources\*.*">
       <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
     </None>
   </ItemGroup>
